% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{check_null}
\alias{check_null}
\title{Check if all elements of a db "emi" is NULL.}
\usage{
check_null(db)
}
\arguments{
\item{db}{The DB object to check for NULL}
}
\value{
TRUE if all elements are NULL or FALSE.
}
\description{
This is specific for Dr. Anene's DBs.
It is likely useful after subsetting the DB.
}
