% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{checkN}
\alias{checkN}
\title{Check for Numeric Columns in a Data Frame or Matrix}
\usage{
checkN(df, l)
}
\arguments{
\item{df}{A data frame or matrix to check.}

\item{l}{A character string used to identify the input in error messages.}
}
\value{
No return value; called for side effects. Stops execution if non-numeric
columns are detected.
}
\description{
Validates that all columns in the provided data frame or matrix are numeric.
If any columns are non-numeric, the function stops execution and returns an error
message listing the offending columns.
}
\examples{
\dontrun{
checkN(mtcars, "mtcars") # should pass
checkN(iris, "iris")     # will error because of non-numeric columns
}
}
