% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{compute_mad}
\alias{compute_mad}
\title{Compute median absolute deviation with constant.}
\usage{
compute_mad(x)
}
\arguments{
\item{x}{A vector of values to compute.}

\item{constant}{The constant value to add mad equal to 0.}

\item{na.rm}{Option to remove "TRUE" or not "FALSE" NA values.}
}
\value{
MAD for each value in the vector
}
\description{
Compute median absolute deviation with constant.
}
\examples{
compute_mad(c(2, 4, 7, 8,9))
}
