% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TranCi.R
\name{ebayes_prior}
\alias{ebayes_prior}
\title{Estimate Empirical Bayes Prior Parameters}
\usage{
ebayes_prior(x)
}
\arguments{
\item{x}{A numeric matrix of expression values with features as rows and samples as columns.}
}
\value{
A named numeric vector with elements:
\describe{
  \item{\code{s2}}{Prior variance (\eqn{s_0^2})}
  \item{\code{df}}{Prior degrees of freedom (\eqn{d_0})}
}
}
\description{
Computes the prior variance (\eqn{s_0^2}) and prior degrees of freedom (\eqn{d_0})
for a given expression matrix using empirical Bayes shrinkage, as implemented in the \pkg{limma} package.
}
\details{
This function fits a simple intercept-only linear model to the input data and estimates
hyper parameters required for moderated t-statistics.

Works best with log2 transformed matrix log2(x)
}
\examples{
mat <- matrix(rnorm(1000), nrow = 100, ncol = 10)
gprior(mat)

}
