% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.R
\name{eRNASeq}
\alias{eRNASeq}
\title{Extract eRNA sequences from a FASTA file using BED coordinates}
\usage{
eRNASeq(b, fasta)
}
\arguments{
\item{b}{Data frame with intervals. It must "chr", "start", "end", "strand" as names. With any other mcols.}

\item{fasta}{Character. Path to the reference genome in FASTA format.}
}
\value{
A \code{DNAStringSet} object containing the extracted sequences.
}
\description{
This function extracts DNA sequences for eRNAs from a reference FASTA file based on
coordinates provided in a BED file. It optionally supports strand-specific
extraction and writes the result to a new FASTA file.
}
