% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna2rna.R
\name{getGRange}
\alias{getGRange}
\title{Get genomic range object for annGE or chimericPRO.}
\usage{
getGRange(tab = exmp)
}
\arguments{
\item{tab}{}
}
\value{
Returns GRanges
}
\description{
Make GRanges object from dataframe.
This function expects a specific input so make sure you have that.
It must have columns named "chr", "start", "end", "strand"
}
