% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eDecosus.R
\name{GSEAms}
\alias{GSEAms}
\title{Compute ssGSEA profiles and scores for all samples in a matrix for a single gene set}
\usage{
GSEAms(mat, gene.set, alpha = 0.25, norm = TRUE)
}
\arguments{
\item{mat}{A numeric matrix of gene expression data with genes as rows and samples as columns. Row names must be gene identifiers.}

\item{gene.set}{A character vector of gene identifiers representing the gene set of interest.}

\item{alpha}{Exponent parameter used for the weighting of ranks (default: 0.25). Higher values give more weight to highly expressed genes.}

\item{norm}{Logical; if TRUE (default), the enrichment profile is normalized using min-max scaling as in GSVA.}
}
\value{
A list with two elements:
\describe{
  \item{profiles}{A matrix of full ssGSEA enrichment profiles, with one column per sample.}
  \item{scores}{A numeric vector of ssGSEA scores (the final enrichment value per sample).}
}
}
\description{
This function applies single-sample GSEA (ssGSEA) across all columns (samples) in an expression matrix.
It returns JUST final ssGSEA scores. If you need the full profile then run the ssGSEA itslf with apply.
}
\examples{
mat <- matrix(rnorm(1000), nrow = 100)
rownames(mat) <- paste0("Gene", 1:100)
gene.set <- paste0("Gene", sample(1:100, 10))
result <- ssGSEA_matrix(mat, gene.set)
head(result$scores)

}
