% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{listDB}
\alias{listDB}
\title{List all database from the eRNAkit.}
\usage{
listDB()
}
\value{
A character list of all the dbs included
}
\description{
Function to list DB .rds included with the eRNAkit.
}
\examples{
listDB()
}
