% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{log2_count2cpm}
\alias{log2_count2cpm}
\title{Count to log2 count per million "CPM"}
\usage{
log2_count2cpm(df = "data", side = "r:1.c:2")
}
\arguments{
\item{df:}{Data frame}

\item{side:}{Operate on row-1 or column-2}
}
\value{
Data frame of converted counts
}
\description{
Convert count to log2 CPM on matrix
}
\examples{
log2_count2cpm(df, 2)
}
