% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eRNAkitDB.R
\name{make_window}
\alias{make_window}
\title{Create n.bp Genomic Windows or Retain Original Interval for eRNAkitDB$core}
\usage{
make_window(db = eRNAkitDB, size = 100)
}
\arguments{
\item{db}{A list representing the eRNAkit database object. Must contain
\code{core} data.frame with columns \code{chr}, \code{start}, and \code{end}}

\item{size}{An integer indicating the size of the windows "n".}
}
\value{
A \code{data.frame} with columns \code{chrom}, \code{start}, \code{end}, and \code{name}.
If the interval is longer than size, the result contains multiple rows representing the windows.
For shorter intervals, a single row is returned.
}
\description{
Given a genomic interval, this function splits it into non-overlapping windows
of n base pairs if the interval length is greater than n. If the interval
is shorter than or equal to n, it is returned as-is.
}
