% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.R
\name{make_windows}
\alias{make_windows}
\title{Create n.bp Genomic Windows or Retain Original Interval}
\usage{
make_windows(chrom, start, end, name, size = 100)
}
\arguments{
\item{chrom}{A character or numeric value representing the chromosome name or number.}

\item{start}{An integer representing the start coordinate of the genomic interval (1-based).}

\item{end}{An integer representing the end coordinate of the genomic interval (inclusive).}

\item{name}{A character string providing an identifier for the interval.}

\item{size}{An integer indicating the size of the windows "n".}
}
\value{
A \code{data.frame} with columns \code{chrom}, \code{start}, \code{end}, and \code{name}.
If the interval is longer than size, the result contains multiple rows representing the windows.
For shorter intervals, a single row is returned.
}
\description{
Given a genomic interval, this function splits it into non-overlapping windows
of n base pairs if the interval length is greater than n. If the interval
is shorter than or equal to n, it is returned as-is.
}
