% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{max_cols}
\alias{max_cols}
\title{Get id of max column by rows.}
\usage{
max_cols(x)
}
\arguments{
\item{x}{Named vector to get max values.}
}
\value{
Returns the max and second max ids.
}
\description{
Function to get ids of max columns in a vector.
Second max column is include for future tooling.
0 values are set to NA to avoid return 0 as max.
}
\examples{
max_cols(c(A = 10, B = 20, C = 30))
}
