% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{max_cols_matrix}
\alias{max_cols_matrix}
\title{Get id of max column by rows.}
\usage{
max_cols_matrix(df)
}
\arguments{
\item{df}{Data frame or matrix to get max values by row.}
}
\value{
Returns a data frame of max and second max column ids.
}
\description{
Function to get ids of max columns in a data frame by row.
Second max column is include for future tooling.
0 values are set to NA to avoid return 0 as max.
}
\examples{
max_cols(matrix(1:9, nrow = 3, ncol = 3))
}
