% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna2rna.R
\name{parseGR}
\alias{parseGR}
\title{Parse a Genomic Coordinate String into a GRanges Object}
\usage{
parseGR(c)
}
\arguments{
\item{c}{A character string specifying the genomic region.
Valid formats include `"chr1:100-200"`, `"CHR1:100-200"`, or `"1:100-200"`.
Whitespace is ignored. Chromosome prefix is normalized to lowercase `"chr"` format.}
}
\value{
A `GRanges` object corresponding to the parsed genomic range.
}
\description{
Takes a genomic coordinate string in the format `"chr1:100-200"` or `"1:100-200"`
(case-insensitive) and returns a `GRanges` object representing that region.
}
\examples{
parse_genomic_range("chr1:100-200")
parse_genomic_range("CHR2:1000-2000")
parse_genomic_range("3:500-1000")

}
