% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{permutation_outliers}
\alias{permutation_outliers}
\title{Detect outliers in a named vector using permutations.}
\usage{
permutation_outliers(x, n = 1000)
}
\arguments{
\item{x}{A named vector of values.}

\item{n}{Number of permutations to run.}
}
\value{
A table of all the results including intermediate tables.
}
\description{
Function to detect outliers in a named vector of values.
It uses permutations to get empirical p-values.
}
\examples{
permutation_outliers(c(2, 4, 7, 8,9), n=100)
}
