% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotLoc}
\alias{plotLoc}
\title{Plot the figures for localisation data.}
\usage{
plotLoc(df, t = c("p1", "p2", "p3"))
}
\arguments{
\item{df}{Data frame to plot from the winloc list of tables.}

\item{t}{The type of plot p1, p2, or p3, matching the winloc table}
}
\value{
GGplot2 plot
}
\description{
This function expects tables pre-processed by the winloc function.
}
