% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotOC}
\alias{plotOC}
\title{Plot Expression Across Tissue or Cells}
\usage{
plotOC(x, t = "Organ")
}
\arguments{
\item{x}{A data frame containing the expression data with columns representing organs or cells and an `E` column indicating the eRNA expression values.}

\item{t}{A string specifying the column name to be used for the x-axis. Defaults to `"Organ"`. It can be any column name in `x` (e.g., `"Cells"`).}
}
\value{
A `ggplot` object showing the expression data as a bar plot.
}
\description{
This function generates a bar plot showing the expression levels of eRNAs across different organs or cells.
It expects the data from emi db.
The data is transformed into a long format using `pivot_longer`, excluding certain columns (`E`, `Specificity_Score`, `Is_Specific`, `expressed`).
The plot is ordered by expression values in descending order, and a color scale is used based on the `E` column.
}
\examples{
# Example 2: Plot expression across cells
plotOC(x2, "Cells")
}
