% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TranCi.R
\name{process_vcf}
\alias{process_vcf}
\title{Process VCF files from a folder}
\usage{
process_vcf(folder, n = 2)
}
\arguments{
\item{folder}{A character string giving the path to the folder containing `.vcf` files.}

\item{n}{Minimum number of samples in which a variant must appear to be retained. Default is 2.}
}
\value{
A data frame with columns `id` (variant identifier) and `sample` (sample name).
}
\description{
This function reads all `.vcf` files in a given folder, filters for PASS SNPs (single nucleotide variants),
and compiles them into a long-format data frame where each row is a variant found in a sample.
Variants are retained only if they are found in at least `n` samples. This function can handle .gz
}
\examples{
\dontrun{
df <- process_vcfs("path/to/vcf_folder", n = 3)
head(df)
}
}
