% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna2rna.R
\name{R2R}
\alias{R2R}
\title{Extract exact location of R2R and count the eRNA:mRNA interactions.}
\usage{
R2R(chimGR, geneGR, enhGR)
}
\arguments{
\item{chimGR}{Chimeric junction GRanges object. It should be made from ChimPRO output.}

\item{geneGR}{Gene annotation GRanges object. It should be made from eRNAkitDB$G.}

\item{enhGR}{Enhancer annotation GRanges object. It should be made from eRNAkitDB$E.}
}
\value{
Interactions counts. Uses IDs for naming enhancer and genes.
}
\description{
Function to extract and count eRNA:mRNA interactions from GRanges object.
}
