% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{remove_zeros}
\alias{remove_zeros}
\title{Filter matrix to remove rows with sum 0.}
\usage{
remove_zeros(matrix)
}
\arguments{
\item{matrix}{}
}
\value{
The filtered matrix
}
\description{
Filter matrix to remove rows with sum 0.
}
\examples{
remove_zeros(matrix(1:9, nrow = 3, ncol = 3))
}
