% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna2rna.R
\name{rna2rna}
\alias{rna2rna}
\title{Extract and count the eRNA:mRNA interactions.}
\usage{
rna2rna(chimGR, geneGR, enhGR)
}
\arguments{
\item{chimGR}{Chimeric junction GRanges object. It should be made from ChimPRO output.}

\item{geneGR}{Gene annotation GRanges object. It should be made from annGE$gene.}

\item{enhGR}{Enhancer annotation GRanges object. It should be made from annGE$erna.}
}
\value{
Interactions counts. Uses IDs for naming enhancer and genes.
}
\description{
Function to extract and count eRNA:mRNA interactions from GRanges object.
}
