% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna2rna.R
\name{rna2rnaR}
\alias{rna2rnaR}
\title{Extract extract locations of eRNA:mRNA interactions.}
\usage{
rna2rnaR(chimGR, geneGR, enhGR)
}
\arguments{
\item{chimGR}{Chimeric junction GRanges object. It should be made from ChimPRO output.}

\item{geneGR}{Gene annotation GRanges object. It should be made from annGE$gene.}

\item{enhGR}{Enhancer annotation GRanges object. It should be made from annGE$erna.}
}
\value{
bedpe type table. xx.x and xx.y are the target pairs
}
\description{
Function to extract locations of eRNA:mRNA interactions from GRanges object.
}
