% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{subDB}
\alias{subDB}
\title{Filter a list of data frames by column and value}
\usage{
subDB(db, col, value)
}
\arguments{
\item{db}{A list of data frames.}

\item{col}{The name of the column to filter by (e.g., "E" or "G").}

\item{value}{The value to filter for.}
}
\value{
A subset of the db
}
\description{
This function filters each data frame in a list by a specific value in a specified column.
It skips any data frames that do not contain the column.
}
\examples{
subDB(emi, "E", "en100035")
}
