% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cisR.R
\name{to_dete}
\alias{to_dete}
\title{Link a given eRNA to significant impact on mRNA decay or translation efficiency}
\usage{
to_dete(E, DB, t = 0.1)
}
\arguments{
\item{E}{A vector containing eRNAs of interest. eRNA IDs must match the naming in eRNAkitDB.}

\item{DB}{The eRNAkitDB object.}

\item{t}{Numeric threshold for significance on adjusted p-values (False Discovery Rate). Default is 0.1.}
}
\value{
A named list with two elements:
\itemize{
  \item \code{de}: data.frame of significant mRNA decay impact.
  \item \code{te}: data.frame of significant translational efficiency impact.
}
}
\description{
This function processes the "eDecay2mRNA" and "TE2mRNA" slots on eRNAkitDB to get their impact on target mRNA.
It requires the "G" object in the DB map the gene symbol of those slots to the Ensemble id.
}
\details{
This expects the exact slot naming convention from the eRNAkitDB.
The output retains the columns in the eRNAkitDB slots but replaces pair for easy mapping.
}
\examples{
\dontrun{
results <- to_dete(E = eRNAs,
                  DB = eRNAkitDB,
                  t = 0.1)
head(results$de)
head(results$te)
}

}
