% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cisR.R
\name{var2erna}
\alias{var2erna}
\title{Map Variants to Overlapping eRNAs}
\usage{
var2erna(output, eRNA = "db")
}
\arguments{
\item{output}{A data.frame with a column named `id` in the format "chr:pos:ref:alt".
Typically the output of \code{\link{process_vcfs}}.}

\item{eRNA}{eRNAkitDB core object.}
}
\value{
A data.frame with columns \code{id} (variant ID) and \code{E} (eRNA ID).
Returns \code{NULL} if no overlaps are found.
}
\description{
This function takes a set of variants (from process_vcf() function)
and returns overlaps with enhancer RNAs (eRNAs) using genomic coordinates.
}
\examples{
\dontrun{
v <- process_vcf("vcf_folder/")
hits <- var2erna(v)
head(hits)
}
}
