% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{wide2long}
\alias{wide2long}
\title{Pivot a wide db table to long}
\usage{
wide2long(df, keep_cols)
}
\arguments{
\item{df}{A data frame in wide format.}

\item{keep_cols}{A character vector of column names to retain (not pivot).}
}
\value{
A data frame in long format with columns: the retained identifiers,
        a `variable` column for the former column names, and a `value` column for the values.
}
\description{
Converts wide emi db table to long format for processing
}
