% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{winLoc}
\alias{winLoc}
\title{Generate location-based subsets from emi loc table}
\usage{
winLoc(subloc)
}
\arguments{
\item{subloc}{emi$loc subset prefered.}
}
\value{
A named list with components: `p1`, `p2`, `p3`, and `d`.
}
\description{
This function processesthe emi db loc table.
}
\details{
It returns:
- p1: CPM values split by RNA type (only cytosol and nucleus)
- p2: REI values for `cytosol` polyA+ RNA split by Cell line
- p3: REI values comparing subcellular cytoplasmic compartments to cytosol (total RNA), split by Enhancer (E)
- d:  The original filtered input dataframe
}
