% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna2rna.R
\name{byRange}
\alias{byRange}
\title{Find Overlaps Between emi$core and a Genomic Coordinate String}
\usage{
byRange(erob, s)
}
\arguments{
\item{erob}{A `GRanges` object to search within.}

\item{s}{A character string in the format "chr1:100-200" or "1:100-200".}
}
\value{
A data.frame of overlapping GRanges entries with metadata columns.
Returns an empty data.frame if no overlaps are found.
}
\description{
Parses a coordinate string and finds overlapping entries in a emi$core.
Returns a data.frame including all metadata columns.
}
\examples{
find_overlap_df(my_gr, "chr1:2000-3000")

}
