% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna2rna.R
\name{chimericPRO}
\alias{chimericPRO}
\title{Process chimeric junction file from star-alignment.}
\usage{
chimericPRO(file)
}
\arguments{
\item{file}{The file path for the chimeric.junction from star.}
}
\value{
A list with n1 and n2 pairs. The name column match the ID.
}
\description{
Function to process chimeric junction files into a bedpe format.
It splits the pairs to make it easy to find overlaps.
}
