% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{eeCigar}
\alias{eeCigar}
\title{Extract end coordinates from cigar string.}
\usage{
eeCigar(start = 2, strand = "+", cigar = "10M5D20M2I")
}
\arguments{
\item{start}{Start position.}

\item{strand}{The strand. Defualt is + standard.}

\item{cigar}{The cigar string.}
}
\value{
The end coordinate.
}
\description{
Function to extract end position from cigar string.
The function only adds M, D and S to the count.
}
