% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TranCi.R
\name{link2rr}
\alias{link2rr}
\title{Link eRNAs to their Regulatory Targets}
\usage{
link2rr(output2, R2R = "dd")
}
\arguments{
\item{output2}{A data frame with a column `E` containing eRNA IDs. Out of var2erna}

\item{R2R}{eRNAkitDB R2R object.}
}
\value{
A data frame with filtered R2R interactions and added `"interaction"` column set to `"R2R"`.
}
\description{
This function filters the R2R interaction table to return only entries
matching a given set of eRNAs (from `output2`).
}
\examples{
\dontrun{
output2 <- var2erna(output)
result <- linkE2RD(output2)
}
}
