% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{loadDB}
\alias{loadDB}
\title{Load a database from the eRNAkit.}
\usage{
loadDB(name)
}
\arguments{
\item{name}{Name of the db set to load (without .rds)}
}
\value{
The loaded R object
}
\description{
Function to load DB .rds DB into environment.
}
\examples{
ribo <- loadDB("ribo")
}
