% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cisR.R
\name{TranCi}
\alias{TranCi}
\title{Prioritize Functional Regulatory Variants Using eRNA–mRNA Interactions.}
\usage{
TranCi(
  vcf = "path/to/vcf",
  E = "eRNA",
  G = "mRNA",
  n = 2,
  fdr = 0.1,
  db = eRNAkitDB
)
}
\arguments{
\item{vcf}{Path to a VCF file containing variant calls. The file names must match the column names in E and G}

\item{E}{Data frame of eRNA expression. Row names must match the eRNA IDs from eRNAkitDB and column names must match vcf file names.}

\item{G}{Data frame of eRNA expression. Row names must match the Ensemble gene IDs and column names must match vcf file names.}

\item{n}{Minimum number of samples to contain the a variant for it to be tested. (default: `2`).}

\item{fdr}{False discovery rate threshold for filtering results (default: `0.1`).}

\item{db}{eRNAkitDB object.}
}
\value{
A list with the following components:
\describe{
  \item{var}{Preprocessed variant data.}
  \item{var2erna}{Mapped variants to eRNAs.}
  \item{einteraction}{eRNA-mRNA interactions involving those variants.}
  \item{varReg}{Final regulatory variant prioritization results. With hit - significant hit, E and G full test for checks}
}
}
\description{
This function identifies and prioritizes putative functional regulatory non-coding variants
by integrating variant data with enhancer RNA (eRNA) to mRNA interaction networks.
It processes a VCF file, maps variants to eRNAs, identifies downstream regulatory targets,
and links variants to potential gene regulatory effects using the `eRNAkit` database.
}
