% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TranCi.R
\name{TranCi_explore}
\alias{TranCi_explore}
\title{Identify Potential Regulatory Variants}
\usage{
TranCi_explore(input, fdr = 0.1, db = eRNAkitDB)
}
\arguments{
\item{input}{Data frame containing variants. Must have exactly three columns: `chr`,
`start`, and `end`.}

\item{fdr}{Numeric. FDR threshold for significance (default is 0.1).}

\item{db}{eRNAkitDB object.}
}
\value{
A named list with result objects.
}
\description{
This function predicts potential regulatory non-coding variants based on target mRNA decay and translation efficiency (TE).
It is tailored for exploratory use in **single-sample scenarios**, such as variant interpretation, rather than full expression-based
prioritisation across a cohort.
}
\details{
The input `input` must be a data frame containing only the variants. It must include
**three columns** named **chr**, **start**, and **end**. Start and end must be the same value.
}
\examples{
# Assuming input_df has columns chr, start, end
# and eRNAkitDB is loaded with required tables:
result <- to_dete(input_df, fdr = 0.05, db = eRNAkitDB)

}
